Prototyping GUI (#85)
Offers a generic GUI that can be used with any game (class gui.PrototypeGUI.java). This offers basic visualisation of all components in the game (package gui.views includes the default displays of all component types), game state information and a simple list of buttons of all available actions.

The components in the main view can also be dragged around (short click + drag -> child components, long click + drag -> collections of components, i.e. decks and areas), with multi-select available as well. A 'search' button allows for observing the contents of a selected deck.

Tested in ExplodingKittens and TicTacToe.

Workflow would be using this directly when designing the game to be able to immediately see if rules/actions/state works appropriately, then making a copy of this GUI + component views classes in the game's package and modifying these for the particular needs of the game (e.g. including images, better distribution of where components are displayed, better user interaction etc.).

For PO, gui should be auotmatically displaying what the active player sees if PARTIAL_OBSERVABLE constant in core.CoreConstants.java is set to true.

Further possible developments:

- More spreadout display of components to start with
- Saving initial component locations to file + reading from file
- Components removed from decks when dragged out, and put back when on top of decks, for rearranging the game space manually (and again option to save configuration)
- More easy customisation of component display
- Merging functionality with the game flow diagram display and adding GUI modification of game rules