Fixed bug in copy(id) for ColtExpress, plus IStatisticLogger interface
When copying a ColtExpress game state from the perspective of a player, it was possible for Bullet cards to end up in the plannedActions pile (where, for example, the card was played while in a tunnel). This then causes an IndexOutOfBoundsError later as a Neutral Bullet card has playerId of -1 when it is processed...

When randomising the unknown planned action cards, we now exclude all Bullet cards, as they are invalid to play as actions.