Exploding Kittens bug fix
There was an edge case when a Favor action had a target of another player with an EmptyDeck. This could happen if the target had cards at the point at which the Favor was played, but not after all reaction cards (such as a double Nope). In this case no actions were calculated by the forwardModel.
To fix this a GiveCard action is now possible with an index of -1, to indicate there are no cards to Give. This is only made available as an action if the target has an empty deck.