Added getGameScore() to AbstractGameState.
The idea here is to provide a general method to find the underlying game score, outside of the win/lose state.
This does not apply to all games (ExplodingKittens or Pandemic for example), in which there is a single victory/game end condition.

It does apply to other games with a well-defined 'score'.