Some refactoring and improvements to reporting abilities.
GameReportII can now accept a number of different Listeners and Loggers. So far this works for GameReportListener which produces general game level attributes, and the new DominionListener which gathers Dominion specific data. The latter in particular is a nice template to use for future Listeners to gather bulk data of interest.