Added Lee's Algorithm to work out exact distance between Figures
Lee's Algorithm allows a more precise distance between two coordinates on the board, which means the Heuristics can more easily account for if a Hero is close to a Monster or not by how easily they can move to it.

The previous method of purely checking the coordinates did not account for if there were walls in the way, commonly leading the Heroes to get stuck in a corner that was unable to get any closer as a result.