Added GameScore
Admittedly the scoring is a little arbitrary, but the Players' scores are determined by how close each victory condition is to being achieved.

For Acolyte of Saradyn, the score is determined by how many Barghests are defeated, how many players are currently defeated, and how much Fatigue the Overlord has.

The maximum score attainable is either
- (Heroes) Defeat all the Barghests with 0 Fatigue suffered, with no Heroes defeated
- (Overlord) Defeat all the Heroes whilst scoring the required Fatigue, with no Barghests defeated