(Hopefully) Fixed an infinite loop with Jain's Hero Ability
Jain would keep trying to convert damage to Fatigue on an attack that she was registered as the attacker for, rather than the defender.

This would lead to an infinite loop of a Monster trying to attack her, Jain requesting to reduce the damage, and then the attack not registering at all, causing the Monster to attempt the attack again, and repeat ad nauseum.

This does explain why Apocrita got stuck on an OSLA runthrough on the very first game for 6 hours for what should have taken only 5 minutes, at least...