Fixed issue with MonsterAbilities and DescentGridBoard
I have no idea why this broke, it was working perfectly fine last night, then all of a sudden it decides the dgs.getMasterBoard() function no longer exists (which is not the case).

No other code was changed so I can only assume this was a hiccup with the branch merging.