Stopped Figures trying to use the next Figure's Surges
The acting figure would change even if an action was currently in progress, which meant that sometimes a Figure could end up using the next Figure's Surges on an in-process attack, rather than using their own.

This is most notable when Goblin Minion 4 tries to use the Barghest Master's Damage +2 Surge, which it should not have access to.