Widow Tarha's Heroic Feat now attacks closest target first
This previously caused an issue where the attack would be cut off because it 'missed' the first target, even though the roll given would have hit the second target.

Now, the closest target is always the first in the pair, such that if the attack misses the first target due to being out of range, it will always miss the second target too.