Feature to allow selecting what criterion increments rollout depth
With this rolloutIncrementType MCTS parameter, it is now possible to
allow the rollout to last for a certain number of turns, or a certain
number of game rounds. This can be helpful when those are more
meaningful metrics of game progression, especially when the number of
moves per turn or per round is not always consistent.